+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-23T04:20:02.873Z",
+    "2026-04-23T06:55:30.818Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {