+++ packages/orq-rc/RELEASES.md
@@ -8569,3 +8569,13 @@ Based on:
 - [typescript v4.8.0-rc.50] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.50 - packages/orq-rc
+
+## 2026-04-23 06:58:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.51] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.51 - packages/orq-rc